requirements.txt
Flask
flask-cors
requests
sentence-transformers
faiss-cpu
ollama
